Added some file structure
